+++ __tests__/utils/version.test.js
@@ -1,9 +1,10 @@
 const axios = require('axios');
-const { sendNotifications } = require('../../notifications/index.js');
-const { initVersionCheck } = require('../../utils/version');
+const { sendNotifications } = require('../../notifications');
+const { init_version_check } = require('../../helpers');
+
 
 jest.mock('axios');
-jest.mock('../../notifications/index.js');
+jest.mock('../../notifications');
 
 const tick = () => new Promise(resolve => jest.requireActual('timers').setImmediate(resolve));
 const packageJson = { version: '1.0.0' };
@@ -12,11 +13,16 @@ jest.mock('../../package.json', () => (packageJson), { virtual: true });
 describe('utils', () => {
     beforeEach(() => {
         jest.resetAllMocks();
-        axios.get.mockResolvedValue({ data: { tag_name: `v1.0.0`, html_url: 'https://github.com/deezy-inc/sat-hunter/releases/tag/v1.0.0' } });
+        axios.get.mockResolvedValue({
+            data: {
+                tag_name: `v1.0.0`,
+                html_url: 'https://github.com/deezy-inc/sat-hunter/releases/tag/v1.0.0'
+            }
+        });
     });
 
     describe('version', () => {
-        describe('initVersionCheck', () => {
+        describe('init_version_check', () => {
             describe('Notifications', () => {
                 test('should send notification about update when local version is different from remote', async () => {
                     // Given
@@ -26,7 +32,7 @@ describe('utils', () => {
                     axios.get.mockResolvedValue({ data: { tag_name: `v${remoteVersion}`, html_url: remoteVersionReleaseUrl } });
 
                     // When
-                    await initVersionCheck();
+                    await init_version_check();
 
                     // Then
                     expect(sendNotifications).toHaveBeenCalledWith(`New version available 1.0.0 -> ${remoteVersion}, check it out at: \n${remoteVersionReleaseUrl}`, 'version_check');
@@ -40,7 +46,7 @@ describe('utils', () => {
                     axios.get.mockResolvedValue({ data: { tag_name: `v${remoteVersion}`, html_url: remoteVersionReleaseUrl } });
 
                     // When
-                    await initVersionCheck();
+                    await init_version_check();
 
                     // Then
                     expect(sendNotifications).toHaveBeenCalledWith(`You are using the latest version! ${remoteVersion}`, 'version_check');
@@ -51,7 +57,7 @@ describe('utils', () => {
                     axios.get.mockRejectedValue();
 
                     // When
-                    await initVersionCheck();
+                    await init_version_check();
 
                     // Then
                     expect(sendNotifications).toHaveBeenCalledWith(`Could not check for new version`, 'version_check');
@@ -69,7 +75,7 @@ describe('utils', () => {
 
                 test('should send notification every 24 hours (default)', async () => {
                     // Startup
-                    await initVersionCheck();
+                    await init_version_check();
                     expect(sendNotifications).toHaveBeenCalledTimes(1);
 
                     // After 24 hours
@@ -86,7 +92,7 @@ describe('utils', () => {
                 test('should send notification every hour (user defined)', async () => {
                     // Startup
                     const oneHour = 1000 * 60 * 60;
-                    await initVersionCheck(oneHour);
+                    await init_version_check(oneHour);
                     expect(sendNotifications).toHaveBeenCalledTimes(1);
 
                     // After 1 hour
